requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
httpx==0.28.1
pytesseract==0.3.13
Pillow==11.1.0
pymupdf==1.25.3
python-dotenv==1.0.1
jinja2==3.1.5
